refactor(utils): removed unused utils + moved types
